pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch]

[project]
name = "visiomode"
description = "An open-source platform for touchscreen-based visuomotor tasks in rodents."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Constantinos Eleftheriou", email = "Constantinos.Eleftheriou@ed.ac.uk" },
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Framework :: Flask",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
  "Topic :: Software Development :: Libraries :: pygame",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.9",
]
keywords = ["touchscreen", "rodent", "behaviour", "neuroscience"]
dependencies = [
  "Flask>=3.0",
  "PyYAML>=6.0",
  "Werkzeug>=3.0.3",
  "pyserial>=3.5",
  "numpy>=1.22",
  "pygame>=2.5.1",
  "pynwb>=2.2.0",
]
dynamic = ["version"]

[project.urls]
Website = "https://www.visiomode.org"
Documentation = "https:/docs.visiomode.org"
Issues = "https://github.com/DuguidLab/visiomode/issues"
Source = "https://github.com/DuguidLab/visiomode"

[project.scripts]
visiomode = "visiomode:run"

[tool.hatch.build]
artifacts = ["*.png"]

[tool.hatch.build.targets.wheel]
packages = ["src/visiomode"]

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.version]
path = "src/visiomode/__about__.py"

[tool.hatch.envs.default]
python = "3.9"

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings-python",
  "pymdown-extensions",
  "material-plausible-plugin",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build --clean --strict"
serve = "mkdocs serve --dev-addr localhost:8000"
deploy = "mkdocs gh-deploy --force"


[tool.hatch.envs.test]
dependencies = [
  "coverage[toml]",
  "pytest",
  "pytest-cov",
  "pytest-mock",
]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.test.scripts]
run-coverage = "pytest --cov-config=pyproject.toml --cov=visiomode --cov=tests --cov-report term --cov-report xml:coverage.xml"
run-coverage-ci = "run-coverage --cov-report=xml:coverage.xml"
run = "run-coverage --no-cov"

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/visiomode tests}"
style = [
  "ruff check {args:src/visiomode tests}",
]
fmt = [
  "ruff check --fix {args:src/visiomode tests}",
  "ruff format {args:src/visiomode tests}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.ruff]
target-version = "py39"
line-length = 120
lint.select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "ISC001",
  # Ignore binding to all interfaces
  "S104",
  # Ignore timezone guff
  "DTZ005"
]
# unfixable = [
#   # Don't touch unused imports
#   "F401",
# ]

[tool.ruff.lint.isort]
known-first-party = ["visiomode"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, relative and unused imports, and unused arguments
"tests/**/*" = ["PLR2004", "S101", "TID252", "F401", "ARG001"]
